Move validateFiles to separate file
